app/requirements.txt
annotated-types==0.6.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
docstring_parser==0.16
Flask==3.0.3
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.46.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==22.0.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.3
kfp==2.7.0
kfp-pipeline-spec==0.3.0
kfp-server-api==2.0.5
kubernetes==26.1.0
lightgbm==4.3.0
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.13.0
shapely==2.0.3
six==1.16.0
tabulate==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
websocket-client==1.7.0
Werkzeug==3.0.2
